Add `mrb_reallocf()` and `mrb_reallocf_simple()`
Commit 9cdf439 changed the behavior of `mrb_realloc()`, but it reverts
this behavior.
Instead, add the `mrb_reallocf()` function that behaves like commit
9cdf439.

Background:

There is a problem if the `mrb_realloc()` function does the heap
freeing.

For example if released when expanding a string object:
- Double free when destroying an object.
- If the same memory address is allocated by later another place, it
  will be a memory corruption.

To fix this problem, it have to temporarily clear the pointer held by
the string object and set its length to 0 before calling `mrb_realloc()`.
This additional work is undesirable.