More stricter `ttype` parameter in `mrb_obj_alloc()`
The purpose of this patch is to make it so that some type tags specified in the `ttype` parameter will fail if they are unrelated to the instance type tag in the `cls` parameter.

***Compatibility Note***

When users call `mrb_obj_alloc()`, giving a type tag unrelated to the class to the `ttype` parameter now raises a `TypeError` exception.
If users intentionally want it, they must pass `NULL` as the `cls` parameter and modify the class of the returned object.